refactor(test): split resources/dispatch/spawner/npc_respawn tests under cap
Behavior-preserving test move (Tier 3 of #529). Each oversized tests.rs
(>700-line hard cap) becomes a tests/ directory of theme-grouped sibling
files, all under cap. No test logic changed; #[test] counts preserved.

- base/resources/tests.rs (779) -> tests/{inventory_slots,overrides}.rs
- base/dispatch/tests.rs (741) -> tests/{routing_logging,chat_speaker_flags}.rs
- cell/spawner/tests.rs (718) -> tests/{spawn_records,live_db_loaders}.rs
- cell/service/ticks/npc_respawn/tests.rs (707) ->
  tests/{fixtures,respawn_lifecycle,respawn_witness_loot}.rs
