+++ crates/services/src/base/resources/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Tests for the `resources` module, split by theme (issue #529):
+//! - [`inventory_slots`]: bag-slot helpers + `pick_first_open_bag`.
+//! - [`overrides`]: cooked-data override application + metadata bumps.
+
+mod inventory_slots;
+mod overrides;
